first part (post excursion and duty) of data file
